requirements.txt
fastapi
uvicorn
chromadb
openai
pypdf
python-dotenv
python-multipart